requirements64.txt
# STOM V2.00 64비트 라이브러리 의존성 파일

numpy==1.26.4
pandas==3.0.0
numba==0.63.1
llvmlite==0.46.0

PyQt5==5.15.11
PyQtWebEngine==5.15.7
pyqtgraph==0.14.0
matplotlib==3.10.8

python-binance==1.0.34
pyupbit==0.2.34
yfinance==1.1.0

python-telegram-bot==22.4
exchange_calendars==4.13.1
beautifulsoup4==4.14.3
requests==2.32.5
websockets==16.0
cmaes==0.12.0
optuna==4.7.0
optuna-dashboard==0.20.0

scikit-learn==1.5.2
lightgbm==4.6.0

psutil==7.2.2
pywin32==311
python-dateutil==2.9.0.post0
pytz==2025.2
tzlocal==5.3.1
loguru==0.7.3
Cython==3.2.4
cryptography==46.0.5
fake-useragent==2.2.0
ntplib==0.4.0
pillow==12.1.0
pyttsx3==2.99
pyzmq==27.1.0
squarify==0.4.4
